Add password expiration plumbing
* Set aggregated expiration values with DPM
* Fix min/max logic when aggregating, and fix unit test
* Add expiration tests when checking if policies are active
* Add expire-password to uses-policies set
* Handle password refresh (clear notifications and sec. holds)
* Handle password expiration (warning and/or wipe synced data)
* Unit tests for provider-level methods
* Refactor common security notification logic
* Placeholder notification strings (need final)

Bug: 3197935

+++ src/com/android/email/provider/AttachmentProvider.java
@@ -482,13 +482,14 @@ public class AttachmentProvider extends ContentProvider {
     }
 
     /**
-     * In support of deleting an account, delete all related attachments.
+     * In support of deleting or wiping an account, delete all related attachments.
      *
      * @param context
-     * @param accountId the account for the mailbox
+     * @param accountId the account to scrub
      */
     public static void deleteAllAccountAttachmentFiles(Context context, long accountId) {
         File[] files = getAttachmentDirectory(context, accountId).listFiles();
+        if (files == null) return;
         for (File file : files) {
             boolean result = file.delete();
             if (!result) {
